Now lazy-loading s3 connection and s3 buckets.
Making these newtork calls proactively in our constructors is not a good
practice. Constructors should only configure the object for use and NOT
actually instantiate non-config object properties.

Perfect case for lazy-loading is testing. Unless the object is heavily
mocked these network calls would 1) require a network connection for
testing even if the object is only instanted in the test and not used
and 2) slow down tests due to network calls.